+++ extras/AFC_unit.py
@@ -42,6 +42,7 @@ class afcUnit:
         self.short_moves_speed  = config.getfloat("short_moves_speed",  self.AFC.short_moves_speed) # Speed in mm/s to move filament when doing short moves. Setting value here overrides values set in AFC.cfg file
         self.short_moves_accel  = config.getfloat("short_moves_accel",  self.AFC.short_moves_accel) # Acceleration in mm/s squared when doing short moves. Setting value here overrides values set in AFC.cfg file
         self.short_move_dis     = config.getfloat("short_move_dis",  self.AFC.short_move_dis)       # Move distance in mm for failsafe moves. Setting value here overrides values set in AFC.cfg file
+        self.max_move_dis       = config.getfloat("max_move_dis", self.AFC.max_move_dis)            # Maximum distance to move filament. AFC breaks filament moves over this number into multiple moves. Useful to lower this number if running into timer too close errors when doing long filament moves. Setting value here overrides values set in AFC.cfg file
 
     def handle_connect(self):
         """
@@ -106,8 +107,7 @@ class afcUnit:
         for the selected unit. Provides buttons for lane calibration, Bowden length calibration, and a back option.
 
         Usage:`UNIT_CALIBRATION UNIT=<unit>`
-        Examples:
-            - `UNIT_CALIBRATION UNIT=Turtle_1`
+        Example: `UNIT_CALIBRATION UNIT=Turtle_1`
         Args:
             None
 
@@ -133,8 +133,7 @@ class afcUnit:
         for each lane, grouped in sets of two, and allows calibration for all lanes or individual lanes.
 
         Usage:`UNIT_LANE_CALIBRATION UNIT=<unit>`
-        Examples:
-            - `UNIT_LANE_CALIBRATION UNIT=Turtle_1`
+        Example: `UNIT_LANE_CALIBRATION UNIT=Turtle_1`
 
         Args:
             UNIT: Specifies the unit to be used in calibration
@@ -175,8 +174,7 @@ class afcUnit:
         for each lane, with a note to only calibrate one lane per unit.
 
         Usage:`UNIT_CALIBRATION UNIT=<unit>`
-        Examples:
-            - `UNIT_CALIBRATION UNIT=Turtle_1`
+        Example: `UNIT_CALIBRATION UNIT=Turtle_1`
 
         Args:
             UNIT: Specifies the unit to be used in calibration